产品详情页查询1.根据产品id查单个产品信息2.根据产品id查多个用户投资记录

+++ micr-dataservice/src/main/java/com/lwm/dataservice/mapper/ProductInfoMapper.java
@@ -14,13 +14,13 @@ import java.util.List;
 */
 public interface ProductInfoMapper {
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(ProductInfo record);
 
     int insertSelective(ProductInfo record);
 
-    ProductInfo selectByPrimaryKey(Long id);
+    ProductInfo selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(ProductInfo record);
